test: Do not expect k8sd to be active after node removal
When a node is removed from a cluster, microcluster removes the k8sd
state folder, which includes the `daemon.yaml` file required for the
service to start. Because of this, the service is in a constant loop
of starting and crashing. We cannot expect the service to be running
in this case.