[Backport 1.32] Cilium version bump & CNCF conformance (#1360)
* build: adds cilium 1.17.1 rocks (#1086)

* Report IPv4 and IPv6 node IPs

This is required for cilium 1.17.1 for dualstack scenarios.

* Adds cilium 1.17.1 rocks

The 1.17.1 cilium rocks have been built, and now we're adding it to
k8s-snap. Note that this version requires the nodes (kubelet) to also
report their IPv4 and IPv6 addresses for dualstack scenarios. k8sd
should be adding both of them to kubelet's --node-ip argument.

Updates gateway-api chart to 1.2.0.

* cilium: Enables cilium session affinity (#1146)

Session affinity is disabled by default in the cilium helm chart. Note
that there are a few session affinity related Conformance tests which
are failing because of this fact:

```
[sig-network] Services should be able to switch session affinity for NodePort service [LinuxOnly] [Conformance]
[sig-network] Services should have session affinity work for service with type clusterIP [LinuxOnly] [Conformance]
[sig-network] Services should have session affinity work for NodePort service [LinuxOnly] [Conformance]
[sig-network] Services should be able to switch session affinity for service with type clusterIP [LinuxOnly] [Conformance]
```

Enabling this should resolve the failures for the tests mentioned above.

Note that the ``sessionAffinity`` helm chart option was introduced in
1.12.0.

* feat: Enables cilium protocol differentiation (#1170)

Starting with cilium 1.17 [1][2], it now supports protocol
differentiation. Setting the loadBalancer.protocolDifferentiation.enabled
will result in ``bpf-lb-proto-diff`` being set in cilium's configmap,
which corresponds to this feature. We need this feature in order for the
following Conformance test to pass:

```
[sig-network] HostPort validates that there is no conflict between pods with same hostPort but different hostIP and protocol [LinuxOnly] [Conformance] [sig-network, Conformance]
```

However, that is not enough for the test to pass. We also need to
support `HostPort`s. According to the documentation [3], because we're
the standard kube-proxy and not cilium's replacement (``kubeProxyReplacement=false``),
we need to also set the Helm chart option ``cni.chainingMode=portmap``.

[1] https://github.com/cilium/cilium/issues/9207
[2] https://github.com/cilium/cilium/pull/33434
[3] https://github.com/cilium/cilium/blob/v1.17.1/Documentation/installation/cni-chaining-portmap.rst#portmap-hostport

* feat: Update the cilium rock versions (#1325)

We have recently reduced the image size of the cilium rock
from 1.47 GB to 811 MB. This should reduce the amount of time
spent by the Kubernetes nodes to become Ready as well.

* use main branch in rock repo

---------

Co-authored-by: Claudiu Belu <claudiu.belu@canonical.com>