fix: Treat unbootstrapped node as a control-plane in the inspection report
In the inspection report, we're collecting various logs for the
Kubernetes services, based on whether the node is a control-plane or
not. However, k8s local-node-status has the exit code 1 if the node
is not bootstrapped / joined, and the node is treated as a worker node.

The issue is that, if a node was a control-plane node that was removed,
we'd still want to collect control-plane related logs.