fix: Properly parse IPv6 endpoints for k8s-apiserver-proxy.json (#1370)
Currently, in scenarios with clusters with IPv6, we will end up with
improper IPv6 addresses in the k8s-apiserver-proxy.json, causing the
k8s-apiserver-proxy service to always fail with the following error:

"Failed to start" err="failed to parse endpoint \"fd42:6db9:e49a:486b:216:3eff:fe48:5318:6443\": address fd42:6db9:e49a:486b:216:3eff:fe48:5318:6443: too many colons in address"

According to the documentation [1], the expectation is that the IPv6
address to be enclosed in brackets ([IPv6]:port).

We're writing the k8s-apiserver-proxy.json in 2 cases:
- when joining a worker node.
- when a new control plane node joined the cluster.

In the first scenario, we are properly handling the IPv6 addresses,
but not in the second. This commit addresses this issue.

[1] https://pkg.go.dev/net#SplitHostPort

(cherry picked from commit 9a9ebf5d950f8ff5e462570e756345e141cc10ef)