fix: Implement Revision for pebble
The Upgrade.k8sd.io CRD has the following regex for its name:

[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*

However, if snap.Revision() would return an empty string, we'd create
an resource with the name "cluster-upgrade-to-rev-", which is invalid.

For pebble, we are returning "", and pebble is being used in
cluster-api-k8s, which results in the following error:

[k8sd] E0507 145 app/hooks_join.go:273] "Failed to handle rollout-upgrade" err="failed to create upgrade: Upgrade.k8sd.io \"cluster-upgrade-to-rev-\" is invalid: metadata.name: Invalid value: \"cluster-upgrade-to-rev-\": a lowercase RFC 1123 subdomain must consist of lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character (e.g. 'example.com', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*')" logger="k8sd" hook="postJoin"

For Pebble, we're now returning the k8s revision instead.