fix: Adds logging when rollout upgrade fails (#1408)
Currently, when a control plane joins, and handleRollOutUpgrade returns
an error, the node is then removed from the cluster. The reason for this
failure is not logged, which makes it harder to debug.

Adds a log line for this scenario.

(cherry picked from commit 4b01cc8300d1fca27cc058b1330f1dc1eaa595e3)