test: Sets default FLAVOR to classic (#1427)
Currently, the test_version_upgrades fails on previous release branches
if TEST_VERSION_UPGRADE_CHANNELS is set to "recent N". The test is
supposed to get a list of snap channels to test the upgrades.

However, by default, config.FLAVOR (TEST_FLAVOR env var), is "", and that
flavor is used by snap.get_most_stable_channels to filter the snap
channels. As it is empty, it creates an invalid regex:

(\d+)\.(\d+)-\/(stable|candidate|beta|edge)

This regex cannot match the snap channels we'd expect (e.g.:
1.32-classic/stable).

Sets config.FLAVOR to classic, as that's the currently supported
version.