fix: Perform PreInitChecks as early as possible (#1423)
Currently, when bootstrapping or joining a node, we are doing a few
pre-init checks: make sure the binaries work, kubernetes-related ports
are not in use, the standard or given containerd path is not currently
in use.

However, it takes a while to get to that stage, while we can easily
check this client-side much faster, immediately giving feedback to the
user that something will go wrong during bootstrap / joining.

Additionally, if the join cluster operation fails because of a pre-init
check server-side, the token will have been used and no longer
available. The user would have to generate another token and try again.
This would not be an issue if we perform these checks client-side.

Doing these checks early will improve the user experience.

(cherry picked from commit 9551cca00f1f7247f59c0a2f0bec89dedf0cec0b)