fix: Don't print k8s binary outputs in CLI (#1546) (#1547)
In #1423, a pre-init check was introduced on the client-side.
In this check, the k8s binaries are started to ensure they do not hang
later. However, the output was printed to stdout causing e.g. the `k8s
bootstrap` command to be polluted with debug messages.

This commit pipes the output into the void instead.

(cherry picked from commit 06e922759417456c89b0729329dbdb24cceab7b3)

Co-authored-by: Benjamin Schimke <benjamin.schimke@canonical.com>