Page layout rework
This page was quite information dense which made it hard to follow. I have tried to reduce the headings, make the wording clearer and group by tabs so the user can progress through the guide smoothly. The tabs are grouped so if the user chooses side-load images, that tab will automatically be selected on the following containerd section