fix: Introduce jitter to reconciliation retries (#1613)
Helm `apply` of the feature reconciliations might get into a deadlock since feature controllers run independently.  For features that share the same underlying chart the situation is even more dire since there are `num_cp_nodes * num_sibling_features` goroutines that try to apply the same Helm chart. Applying the same Helm chart concurrently or in parallel might fail with `another operation is in progress`.
This PR introduces a jitter to the reconciliation retries in order to reduce the concurrent/parallel Helm operations.

---------

Signed-off-by: Homayoon (Hue) Alimohammadi <homayoon.alimohammadi@canonical.com>