feat(upgrade): Wait for worker nodes to get upgraded as well (#1643)
This PR implements a way to wait for the worker nodes in the upgrade process as well, before triggering the feature controllers.

We've introduced a `version.Info` struct that contains all the necessary information about the current version of the installed k8s-snap. The node label controller is responsible for annotating the node with this struct.
The upgrade controller will watch for this annotation and add the node name to the list of `upgradedNodes`.
Once all the nodes (including workers) are available in the `upgradedNodes` list, the upgrade controller will transition the upgrade phase from `NodeUpgrade` to `FeatureUpgrade`, triggering the feature controllers and proceeding with the upgrade.

---------

Signed-off-by: Homayoon (Hue) Alimohammadi <homayoon.alimohammadi@canonical.com>