docs: Add node removal notice (#1697) (#1717)
* Add node removal notice

If you delete the snap before remvonig the node from the cluster, the node hangs and is unreachable in the cluster. There is work planned to fix this, but for now adding a warning to our docs on the importance of removing the node first

* md lint fix

* Apply review suggestions

* Remove broken link

Dqlite link is causing the linkcheck to fail. Temporarily remove this link

(cherry picked from commit 92cdfaadee72447ce890976385cfa0ad7000495f)

Co-authored-by: Niamh Hennigan <niamh.hennigan@canonical.com>