fix: Force removing lost nodes from cluster
The `--force` flag for `k8s remove-node` did not work for lost nodes
because it required a node to be available to remove it.

This commit ensures that all mentions of the leaving node will be
removed (etcd, microcluster, k8s-dqlite, Kubernetes) and not fail if the
node does not exist already in one of them.