test: enable node_join_succeeds_when_original_control_plane_is_down
The test logic/order was slightly modified. The dead node is now removed right
after the killing it instead of trying to join a new node first.
This is because microcluster fails joining a new node if the join token
contains an unavailble node. This is fixed in a separate commit to
microcluster.