docs: Remove linkcheck (#1912) (#1921)
* Remove linkcheck

There are intermitent failures in our CI in relation to the linkcheck test. There has been investigation into the issue and there needs to be work done to make the test more suitable for our needs. For now, I am removing the check and will reinstate when the better solution is reached

* Trivial change to index to trigger docs workflow

* Update comment to fix error

The previous implementation was not fully commented out and this was causing errors

* Fetch depth

(cherry picked from commit 8b70d1fdd66e36ba8dda2d7826fb4681da6ed956)

Co-authored-by: Niamh Hennigan <niamh.hennigan@canonical.com>