Update tabs
The tabs were glitching slightly on this page. When you selected certain tabs, the page would jump up and then scroll down to the selected tab, even if the tab was already on the screen. Changing to the sphinx native tabs rather than sphinx tabs extension seems to fix this issue.