ci: move datastore spread testing to weekly (#2111)
Addition of the datastore to the nightly matrix resulted in github not
being able to load the nightly run page at all. Instead we move the
testing to weekly to run on latest/edge for amd64 and arm64.

(cherry picked from commit 111212fc9e21384265b0d1964f980193238f9e6d)