docs: Apply feedback based on UXR sessions (#2119)
* Apply feedback based on UXR sessions

In the latest 2 user experience research sessions I noticed some common pain points. It took too many clicks to get to the tutorial so I have changed the link to the tutorial tile on the index pages. There was a memory issue with the recommended 4G of RAM so I have increased it across the docs. Reference and Explanation tiles on index pages were also flipped. Some other misc updating of wording to address misunderstandings

* Fix md error